Change FormData#entries polyfill to a wrapper
A wrapper is cleaner at the implementation side as it avoids the need
to mess with patching native constructors and having inconsistent
prototypes between polyfill constructor and native constructor.
