+++ src/polyfills.js
@@ -26,7 +26,6 @@ import {install as installMathSign} from './polyfills/math-sign';
 import {install as installObjectAssign} from './polyfills/object-assign';
 import {install as installPromise} from './polyfills/promise';
 import {install as installArrayIncludes} from './polyfills/array-includes';
-import {install as installFormDataEntries} from './polyfills/formdata-entries';
 import {getMode} from './mode';
 
 /**
@@ -45,4 +44,3 @@ installObjectAssign(self);
 installPromise(self);
 installDocContains(self);
 installArrayIncludes(self);
-installFormDataEntries(self);
